chore: moves auto_complete_name to widgets/misc
