+++ frontend/lib/widgets/movement/add_movement.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:wod_board_app/widgets/auto_complete/auto_complete_name.dart';
 import 'package:wod_board_app/widgets/misc/add_duration_repetition_panel.dart';
+import 'package:wod_board_app/widgets/misc/auto_complete_name.dart';
 
 class AddMovement extends StatefulWidget {
   const AddMovement({super.key});
